Adding more finance implementations and tests

+++ finance/fs/src/main/java/gos/integration/finance/fs/Quote.java
@@ -1,13 +1,15 @@
 package gos.integration.finance.fs;
 
+import java.math.BigDecimal;
+
 public class Quote {
   private String symbol;
-  private float price;
-  private float change;
+  private BigDecimal price;
+  private BigDecimal change;
 
   Quote() {}
 
-  Quote(String symbol, float price, float change) {
+  Quote(String symbol, BigDecimal price, BigDecimal change) {
     this.symbol = symbol;
     this.price = price;
     this.change = change;
@@ -17,11 +19,11 @@ public class Quote {
     return this.symbol;
   }
 
-  public float getPrice() {
+  public BigDecimal getPrice() {
     return this.price;
   }
 
-  public float getChange() {
+  public BigDecimal getChange() {
     return this.change;
   }
 
@@ -29,11 +31,11 @@ public class Quote {
     this.symbol = symbol;
   }
 
-  public void setPrice(float price) {
+  public void setPrice(BigDecimal price) {
     this.price = price;
   }
 
-  public void setChange(float change) {
+  public void setChange(BigDecimal change) {
     this.change = change;
   }
 }
